[refactor] 계정 정보 컴포넌트 분리
